requirements.txt
python-telegram-bot>=20.0
aiohttp>=3.8.0
python-dotenv>=0.19.0
pydantic>=2.0.0
httpx>=0.24.0
python-magic>=0.4.27
aiosqlite>=0.19.0
nltk>=3.8.1
scikit-learn>=1.3.0
